refactoring server.js in small files
