+++ src/utils/validators.ts
@@ -1,4 +1,4 @@
-import { prisma } from "../server.js";
+import { prisma } from "../prisma.js";
 
 const PASSWORD_VALID_REGEX = new RegExp(
   /^(?=.*[A-Z])(?=.*[0-9])(?=.*[a-z]).{6,}$/,
